requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.18.0
alembic==1.13.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
bottle==0.12.25
brotlipy==0.7.0
cachetools==5.3.0
certifi==2021.5.30
charset-normalizer==3.1.0
cmake==3.26.0
colorlog==6.8.2
comm==0.1.2
conda==4.10.3
conda-package-handling==1.7.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.0
fonttools==4.39.0
fqdn==1.5.1
google-auth==2.16.2
google-auth-oauthlib==0.4.6
greenlet==3.0.3
grpcio==1.51.3
tqdm==4.61.2
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server==2.4.0
jupyter-server-fileid==0.8.0
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.6.1
jupyter-ydoc==0.2.3
jupyterlab==3.6.1
jupyterlab-language-pack-zh-cn==3.6.post0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.20.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lit==15.0.7
llvmlite==0.41.1
Mako==1.3.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.3
notebook-shim==0.2.2
numba==0.58.1
numpy==1.24.2
oauthlib==3.2.2
optuna==3.5.0
optuna-dashboard==0.14.0
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pyg-lib==0.3.1+pt20cu118
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel-yaml-conda==0.15.100
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.25
stack-data==0.6.2
supervisor==4.2.5
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
torch @ http://download.pytorch.org/whl/cu118/torch-2.0.0%2Bcu118-cp38-cp38-linux_x86_64.whl
torch-cluster==1.6.3+pt20cu118
torch-geometric==2.4.0
torch-scatter==2.1.2+pt20cu118
torch-sparse==0.6.18+pt20cu118
torch-spline-conv==1.2.2+pt20cu118
torchvision @ http://download.pytorch.org/whl/cu118/torchvision-0.15.1%2Bcu118-cp38-cp38-linux_x86_64.whl
tornado==6.2
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.9.0
uri-template==1.2.0
urllib3=1.26.6
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.15.0
